+++ src/main/java/com/juicycool/backend/domain/stock/service/SellStockService.java
@@ -3,5 +3,5 @@ package com.juicycool.backend.domain.stock.service;
 import com.juicycool.backend.domain.stock.presentation.dto.request.SellStockRequestDto;
 
 public interface SellStockService {
-    void execute(Long stockId, SellStockRequestDto dto);
+    void execute(Integer stockId, SellStockRequestDto dto);
 }
